reserve now + cancel reservation

+++ Implementation/cpmsServer/src/main/java/softwareengineering/manonisgaravattiferretti/cpmsServer/cpHandler/messages/cpmsReq/CancelReservationReq.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class CancelReservationReq {
-    private Integer reservationId;
+    private Long reservationId;
 }
